requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.2
django-heroku==0.3.1
django-picklefield==3.0.1
entrypoints==0.3
grpcio==1.33.2
gunicorn==20.0.4
httplib2==0.18.1
idna==2.7
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
kiwisolver==1.3.0
libcst==0.3.13
llvmlite==0.34.0
lxml==4.6.1
MarkupSafe==1.1.1
mistune==0.8.4
multitasking==0.0.9
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.5
numba==0.51.2
numpy==1.19.2
oauth2client==4.1.3
olefile==0.46
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
plotly==4.11.0
pooch==1.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
proto-plus==1.11.0
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyquery==1.4.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.10.15
requests==2.19.1
resampy==0.2.2
retrying==1.3.3
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
sip==5.5.0
six==1.15.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.50.2
uritemplate==3.0.1
urllib3==1.23
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==4.1.2

